backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
python-dotenv>=1.0,<2.0
openpyxl>=3.1,<4.0
xlrd==1.2.0
mysqlclient>=2.2,<3.0
scikit-learn>=1.4,<2.0
pandas>=2.1,<3.0
numpy>=1.26,<3.0
scipy>=1.11,<2.0
jieba>=0.42,<1.0
